separate noop model
